refactor(student): 長文タブを StoryPanel として切り出す
renderStoryContent とハイライト処理を
src/components/student/StoryPanel.js へ移す。データの取得と生成は
親が持ち、表示だけを受け持たせる。

StudentDashboard は 2,727 → 2,553 行。
